Stopping work on the daml tabs.
Currently we have an unstyled DAML contracts tab and a list of
templates; on the contracts tab we have buttons for each choice in the
template, and these show a form with basically type-appropriate widgets
for the choice when clicked. The form does not currently have a submit
button, however, and so no actions are actually possible on the
contract.

As this work has been de-scoped for now, this will be left on a branch.